Add angular velocity definition
Adds a definition to calculate the angular velocity of the target body around the central axis it revolves.
IE:Mars around the Sun; KSP around the Earths CM; etcetera. There are many ways to calculate this physics
quantity, and indeed many of these definitions may likely become classes as the scripts in this project
begin to build in complexity