Add explanation for v_circ definition
This is to clarify to the user that this instance only really works for velocities whose eccentricities are 0 or
close enough to not factor into signigant figures. For velocities around non-circular bodies the result is a bit
more involved. However, this still relies on the fact that equals areas are swept in equal time.